tests/baseline/protocol-1.2/inspector/console/console-format-collections-expected.txt

CONSOLE MESSAGE: line 15: [object HTMLCollection]
CONSOLE MESSAGE: line 19: [object HTMLCollection]
CONSOLE MESSAGE: line 23: [object HTMLOptionsCollection]
CONSOLE MESSAGE: line 27: [object HTMLAllCollection]
CONSOLE MESSAGE: line 31: [object HTMLFormControlsCollection]
CONSOLE MESSAGE: line 35: [object RadioNodeList]
CONSOLE MESSAGE: line 41: 1,2,
CONSOLE MESSAGE: line 44: [object Object]
CONSOLE MESSAGE: line 51: [object Arguments]
CONSOLE MESSAGE: line 55: [object DOMTokenList]
CONSOLE MESSAGE: line 58: [object Object]
CONSOLE MESSAGE: line 59: [object Object]
CONSOLE MESSAGE: line 61: [object Object]
CONSOLE MESSAGE: line 62: [object Object]
CONSOLE MESSAGE: line 63: [object Object]
CONSOLE MESSAGE: line 64: [object Object]
CONSOLE MESSAGE: line 65: [object Object]
CONSOLE MESSAGE: line 66: [object Object]
CONSOLE MESSAGE: line 15: [object HTMLCollection]
CONSOLE MESSAGE: line 19: [object HTMLCollection]
CONSOLE MESSAGE: line 23: [object HTMLOptionsCollection]
CONSOLE MESSAGE: line 27: [object HTMLAllCollection]
CONSOLE MESSAGE: line 31: [object HTMLFormControlsCollection]
CONSOLE MESSAGE: line 35: [object RadioNodeList]
CONSOLE MESSAGE: line 41: 1,2,
CONSOLE MESSAGE: line 44: [object Object]
CONSOLE MESSAGE: line 51: [object Arguments]
CONSOLE MESSAGE: line 55: [object DOMTokenList]
CONSOLE MESSAGE: line 58: [object Object]
CONSOLE MESSAGE: line 59: [object Object]
CONSOLE MESSAGE: line 61: [object Object]
CONSOLE MESSAGE: line 62: [object Object]
CONSOLE MESSAGE: line 63: [object Object]
CONSOLE MESSAGE: line 64: [object Object]
CONSOLE MESSAGE: line 65: [object Object]
CONSOLE MESSAGE: line 66: [object Object]
Tests that console nicely formats HTML Collections, NodeLists and DOMTokenLists.

console-format-collections.html:15 HTMLCollection[1]
console-format-collections.html:19 HTMLCollection[3]
console-format-collections.html:23 HTMLOptionsCollection[2]
console-format-collections.html:27 HTMLAllCollection[14]
console-format-collections.html:31 HTMLFormControlsCollection[3]
console-format-collections.html:35 RadioNodeList[2]
console-format-collections.html:41 Array[2]
console-format-collections.html:44 NonArrayWithLength
console-format-collections.html:51 Arguments[2]
console-format-collections.html:55 DOMTokenList[3]
console-format-collections.html:58 ArrayLike[5]
console-format-collections.html:59 ArrayLike[4294967295]
console-format-collections.html:61 ArrayLike
console-format-collections.html:62 ArrayLike
console-format-collections.html:63 ArrayLike
console-format-collections.html:64 ArrayLike
console-format-collections.html:65 ArrayLike
console-format-collections.html:66 ArrayLike
console-format-collections.html:15 [select#sel, sel: select#sel]
console-format-collections.html:19 [script, script, script]
console-format-collections.html:23 [option, option, selectedIndex: 0]
console-format-collections.html:27 [html, head, script, script, script, body, p, div.c1.c2.c3, form#f, select#sel, option, option, input, input, f: form#f, sel: select#sel, x: NodeList[2]]
console-format-collections.html:31 [select#sel, input, input, sel: select#sel, x: RadioNodeList[2]]
console-format-collections.html:35 [input, input, value: ""]
console-format-collections.html:41 [1, Array[2]]
console-format-collections.html:44 NonArrayWithLength {keys: Array[0]}
console-format-collections.html:51 [1, "2"]
console-format-collections.html:55 ["c1", "c2", "c3", value: "c1 c2 c3"]
console-format-collections.html:58 [undefined × 5]
console-format-collections.html:59 [undefined × 4294967295]
console-format-collections.html:61 ArrayLike {length: -5}
console-format-collections.html:62 ArrayLike {length: 5.6}
console-format-collections.html:63 ArrayLike {length: NaN}
console-format-collections.html:64 ArrayLike {length: Infinity}
console-format-collections.html:65 ArrayLike {length: -0}
console-format-collections.html:66 ArrayLike {length: 4294967296}
